+++ src/app/app.module.ts
@@ -11,21 +11,27 @@ import {HeaderModule} from "./header/header.module";
 import {ConfigOptionsService} from "./core/service/config-options.service";
 import {CoreModule} from "./core/core.module";
 import {CONSTANTS} from "./core/service/constants.service";
+import {ProductRoutingModule} from "./product/product-routing.module";
+import { PathNotFoundComponent } from './shared/path-not-found/path-not-found.component';
+import { LoginFormComponent } from './shared/login-form/login-form.component';
 
 @NgModule({
   declarations: [
     AppComponent,
+    PathNotFoundComponent,
+    LoginFormComponent,
   ],
   imports: [
     BrowserModule,
     CommonModule,
     FormsModule,
-    AppRoutingModule,
-    ProductModule,
+    CoreModule,
+
     CartModule,
     HeaderModule,
-    CoreModule
+    ProductModule,
 
+    AppRoutingModule,
   ],
   providers: [
     ConfigOptionsService,