+++ src/app/product/product-list/product-list.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Product} from "../model/product";
 import {ProductService} from "../product.service";
 import {CartService} from "../../cart/cart.service";
+import {ActivatedRoute, Router} from "@angular/router";
 
 @Component({
   selector: 'product-list',
@@ -13,7 +14,9 @@ export class ProductListComponent implements OnInit {
   products: Promise<Array<Product>>;
 
   constructor(private productService: ProductService,
-              private cartService: CartService) {
+              private cartService: CartService,
+              private route: ActivatedRoute,
+              private router: Router) {
   }
 
   ngOnInit() {
@@ -26,7 +29,11 @@ export class ProductListComponent implements OnInit {
     }
   }
 
-
+  onView(id: number): void {
+    //this.router.navigate([{ outlets: { productview: ['productview'] } }]);
+    // this.router.navigate([{outlets: {product: 'view'}}]);
+    this.router.navigate([{outlets: {product: [id]}}], {relativeTo: this.route});
+  }
 
 
 }